+++ pac4j-jwt/src/main/java/org/pac4j/jwt/credentials/authenticator/JwtAuthenticator.java
@@ -45,6 +45,9 @@ import static org.pac4j.core.util.CommonHelper.assertNotNull;
 @Setter
 public class JwtAuthenticator extends ProfileDefinitionAware implements Authenticator {
 
+    private static final String NONSIGNED_JWT_ERROR_MSG
+        = "A non-signed JWT cannot be accepted as signature configurations have been defined";
+
     protected final Logger logger = LoggerFactory.getLogger(getClass());
 
     private List<EncryptionConfiguration> encryptionConfigurations = new ArrayList<>();
@@ -53,7 +56,7 @@ public class JwtAuthenticator extends ProfileDefinitionAware implements Authenti
 
     /**
      * Whether encryption is mandatory when encryption configurations are present.
-     * Deprecated: to set to true in v7 or even remove.
+     * Deprecated: to set to true in v7 or even remove?
      * <p>
      * Default is {@code false} for backward compatibility.
      */
@@ -178,14 +181,15 @@ public class JwtAuthenticator extends ProfileDefinitionAware implements Authenti
             var jwt = JWTParser.parse(token);
 
             if (encryptionRequired && !encryptionConfigurations.isEmpty() && !(jwt instanceof EncryptedJWT)) {
-                throw new CredentialsException("A non-encrypted JWT cannot be accepted as encryption configurations have been defined");
+                throw new CredentialsException(
+                    "A non-encrypted JWT cannot be accepted as encryption configurations have been defined and are required");
             }
 
             if (jwt instanceof PlainJWT) {
                 if (signatureConfigurations.isEmpty()) {
                     logger.debug("JWT is not signed and no signature configurations -> verified");
                 } else {
-                    throw new CredentialsException("A non-signed JWT cannot be accepted as signature configurations have been defined");
+                    throw new CredentialsException(NONSIGNED_JWT_ERROR_MSG);
                 }
             } else {
 
@@ -226,7 +230,7 @@ public class JwtAuthenticator extends ProfileDefinitionAware implements Authenti
                 // signed?
                 if (!signatureConfigurations.isEmpty()) {
                     if (signedJWT == null) {
-                        throw new CredentialsException("A non-signed JWT cannot be accepted as signature configurations have been defined");
+                        throw new CredentialsException(NONSIGNED_JWT_ERROR_MSG);
                     }
 
                     logger.debug("JWT is signed");